Removed reference to `get_dict.sh` in the tutorial.
Since there are a handful of errors (w.r.t. to stress numbers) in the
CMU dictionary, I have been also circulating a copy of CMU with the
appropriate fixes. `get_dict.sh` nukes this file, and so it'd be better
if people didn't use it, as per Gretchen's suggestion.