Return of hidden scores file
Revitalizes the scores file. When aligning data,
a hidden file called .scores.csv will be created in the
data directory. (Note this is a name change: it is now a CSV file, so
the name is ".scores.csv", not ".scores.txt", as before).

Re: issue #38, but not quite there yet.