`dictionary.txt` reference removed
There was a stale and confusing reference to `dictionary.txt` in the
README which has now been excised. Rather, the aligner ships with
`eng.dict`, which is a lightly hacked version of the CMU dictionary.
From now on, dictionaries we distribute will be labeled using the
three-character ISO code for said language.

h/t Hyunah Ahn